transliteration_service/requirements.txt
Flask
ai4bharat-transliteration